requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
click==8.1.7
    # via
    #   git-scripts (pyproject.toml)
    #   rich-click
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via git-scripts (pyproject.toml)
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
pygments==2.18.0
    # via rich
rich==13.7.1
    # via rich-click
rich-click==1.8.3
    # via git-scripts (pyproject.toml)
smmap==5.0.1
    # via gitdb
typing-extensions==4.12.2
    # via rich-click
